Less messy messier

+++ src/sst/elements/messier/nvm_request.h
@@ -24,26 +24,27 @@
 #include <sst/core/component.h>
 #include <sst/core/timeConverter.h>
 #include <sst/elements/memHierarchy/memEvent.h>
-#include<map>
-#include<list>
+
+#include <map>
+#include <list>
 
 using namespace SST;
 
-namespace SST{ namespace MessierComponent{
+namespace SST{
+namespace MessierComponent{
 
 class NVM_Request
 {
-
-	public:
-		NVM_Request() {}
-		NVM_Request(long long id, bool R, int size, long long int Add) { req_ID = id; Read = R; Size = size; Address = Add;}
-		long long int req_ID;
-		bool Read;
-		int Size;
-		long long int Address;
-		int meta_data;
-
+    public:
+        NVM_Request() {}
+        NVM_Request(uint64_t id, bool R, int size, uint64_t Add) { req_ID = id; Read = R; Size = size; Address = Add;}
+        uint64_t req_ID;
+        bool Read;
+        int Size;
+        uint64_t Address;
+        int meta_data;
 };
 
-}}
+}
+}
 #endif
